requirements.txt
gpustat
json2html
psutil
pyinstaller
pyyaml
requests
